requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.1.0

# Testing dependencies
pytest == 8.4.2
pytest-asyncio == 1.2.0
httpx == 0.28.1
pytest-cov == 7.0.0